requirements.txt
Django==2.2.6
psycopg2==2.8.3
pytz==2019.2
sqlparse==0.3.0
Pillow==6.2.0
django-excel==0.0.10
pyexcel-xlsx==0.5.7
pyexcel-handsontable==0.0.2
pyexcel-sortable==0.0.2
pyexcel==0.5.15
python-crontab==2.4.0
